chatbot/requirements.txt
streamlit>=1.33
pandas>=2.0
numpy>=1.24
scikit-learn>=1.4
joblib>=1.3
python-dotenv>=1.0
supabase>=2.0
sentence-transformers>=2.7
mutagen>=1.47
google-genai>=1.0.0
